+++ src/main/java/pl/team/carrent/client/ClientController.java
@@ -4,7 +4,6 @@ package pl.team.carrent.client;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
-import pl.team.carrent.service.ClientService;
 
 
 /**
