+++ src/test/java/pl/team/carrent/repository/ClientRepositoryTest.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import pl.team.carrent.client.Client;
+import pl.team.carrent.client.ClientRepository;
 
 import javax.transaction.Transactional;
 import java.util.List;
