+++ src/test/java/pl/team/carrent/repository/InvoiceRepositoryTest.java
@@ -6,6 +6,7 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+import pl.team.carrent.client.ClientRepository;
 import pl.team.carrent.invoice.Invoice;
 import pl.team.carrent.invoice.InvoiceRepository;
