+++ src/test/java/pl/team/carrent/repository/RentHistoryRepositoryTest.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import pl.team.carrent.client.Client;
+import pl.team.carrent.client.ClientRepository;
 import pl.team.carrent.invoice.Invoice;
 import pl.team.carrent.invoice.InvoiceRepository;
 import pl.team.carrent.model.*;
